Move ClassStatisticField enum into new taa.logic.commands.enums package
